% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LobDates.R
\name{as.day}
\alias{as.day}
\title{Pulls out the day as a number}
\usage{
as.day(x, fmat = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{the date}

\item{fmat}{is the date format}
}
\value{
Returns the day
}
\description{
This extracts a day from a date
}
\examples{
x <- "2014-02-14"
as.day(x)
}
