% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LobDates.R
\name{as.month}
\alias{as.month}
\title{Pulls out the month as a number}
\usage{
as.month(x, fmat = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{the date}

\item{fmat}{is the date format}
}
\value{
Returns the month
}
\description{
This extracts a month from a date
}
\examples{
x <- "2014-02-14"
as.month(x)
}
