% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammMeans.R
\name{gammMeans}
\alias{gammMeans}
\title{Pulls out the marginal means from a GAMM built using the gamm4 package}
\usage{
gammMeans(mod, lev, smooth_avg = TRUE)
}
\arguments{
\item{mod}{the gamm4 model}

\item{lev}{is the level for which the means are wanted}

\item{smooth_avg}{sets the smooth term level, TRUE = average for all, FALSE = lowest observed value for all and a number (e.g. 2) can set the level. A vector of numbers can be used if there are multiple smooth terms (e.g. c(2,2))}
}
\value{
Returns a data frame
}
\description{
Returns the marginal means and confidence intervals
}
\examples{
library(gamm4)
library(dplyr)
library(magrittr)
set.seed(88)
dat <- expand.grid(year=factor(2000:2020), month=1:12)
dat \%<>\% mutate(temp=runif(nrow(dat), 50,100))
gm4 <- gamm4(log(temp)~year+s(month), data=dat)
out <- gammMeans(gm4,'year')
out
}
