% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Others.R
\name{numchar}
\alias{numchar}
\title{Change factor to number}
\usage{
numchar(x)
}
\arguments{
\item{x}{the factor}
}
\value{
Returns a number
}
\description{
This Changes a factor to number
}
\examples{
x <- as.factor(88)
numchar(x)
}
