% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Others.R
\name{OpenText}
\alias{OpenText}
\title{Opens a text file and converts to data.frames}
\usage{
OpenText(x)
}
\arguments{
\item{x}{the text file name}
}
\value{
Returns a data.frame
}
\description{
Opens a text file and converts to data.frames
}
\examples{
OpenText(system.file("extdata", "test.txt", package = "LobFunc"))
}
