% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggTheme.R
\name{theme_lob}
\alias{theme_lob}
\title{Change factor to number}
\usage{
theme_lob(Ang = 0, leg = "none")
}
\arguments{
\item{Ang}{the angle of X tick labels}

\item{leg}{should the legend be shown, and if yes, then where. Options include "none", "left", "right","bottom", "top" and probably others.}
}
\value{
sets a theme for ggplots
}
\description{
This Changes a factor to number
}
\examples{
theme_lob(90, "left")
}
