% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Others.R
\name{toXL}
\alias{toXL}
\title{Copy to Clipboard}
\usage{
toXL(x, rnames = FALSE)
}
\arguments{
\item{x}{name of data.frame}

\item{rnames}{logical - should row names be printed}
}
\value{
Clipboard gets loaded up
}
\description{
Tkes a dataframe and copies to to the clipboard so you can paste it into Excel
}
\examples{
dat <- data.frame(year=2000:2021, temperature=23)
toXL(dat)
}
